requirements.txt
Flask==3.1.0
sympy==1.13.3
numpy==2.2.4
matplotlib==3.10.1
openai==1.69.0
python-dotenv==1.0.1
